requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
numpy==1.26.4
scikit-learn==1.5.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
scipy==1.14.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
